pyproject.toml
[tool.poetry]
name = "test-task-pavel"
version = "0.1.0"
description = ""
authors = ["Nikolai Dudinov <nvdudinov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyhamcrest = "^2.1.0"
pytest = "^7.4.3"
requests = "^2.31.0"
pytest-playwright = "^0.4.3"
allure-pytest = "^2.13.2"
black = "^23.10.1"
ruff = "^0.1.3"
faker = "^19.12.0"
page = "^0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "ui_test",
    "api_test"
]
addopts = [
    "--clean-alluredir",
    "--alluredir=allure-results"
]
testpaths = ["tests"]